requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.4.1
filelock==3.8.0
idna==3.3
joblib==1.1.0
lxml==4.9.1
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
searchtweets-v2==1.1.1
six==1.16.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
textblob==0.17.1
tqdm==4.64.1
tweepy==4.10.1
urllib3==1.26.11
